Collapse, keep names.
