--- src/utils/optionsToNotation.ts
@@ -1,11 +1,0 @@
-import type { Notation, RollOptions } from '~types'
-import {
-  formatCoreNotation,
-  formatModifierNotation
-} from './notationFormatters'
-
-export function optionsToNotation<S extends string | number>(
-  options: RollOptions<S>
-): Notation<S> {
-  return `${formatCoreNotation(options)}${formatModifierNotation(options)}` as Notation<S>
-}
